Tratativas de erros adicionadas para cada tipo de função
